requirements.txt
llama-index
langchain
langchain_community
python-dotenv
streamlit
pinecone
streamlit-chat 